requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pyyaml>=6.0
scikit-learn>=1.3
joblib>=1.3
